Create separate test target for the integration tests
