+++ IntegrationTests/IntegrationTests.swift
@@ -1,0 +1,33 @@
+//
+//  IntegrationTests.swift
+//  IntegrationTests
+//
+//  Created by Jacob Persson on 01.03.21.
+//  Copyright © 2021 Zeta Project Gmbh. All rights reserved.
+//
+
+import XCTest
+
+class IntegrationTests: XCTestCase {
+
+    override func setUpWithError() throws {
+        // Put setup code here. This method is called before the invocation of each test method in the class.
+    }
+
+    override func tearDownWithError() throws {
+        // Put teardown code here. This method is called after the invocation of each test method in the class.
+    }
+
+    func testExample() throws {
+        // This is an example of a functional test case.
+        // Use XCTAssert and related functions to verify your tests produce the correct results.
+    }
+
+    func testPerformanceExample() throws {
+        // This is an example of a performance test case.
+        measure {
+            // Put the code you want to measure the time of here.
+        }
+    }
+
+}
