+++ Tests/Source/Integration/AppLockIntegrationTests.swift
@@ -54,60 +54,3 @@ final class AppLockIntegrationTests: IntegrationTest {
     }
 
 }
-
-// MARK: - Mocks
-
-class MockAppLock: AppLockType {
-
-    // MARK: - Metrics
-
-    var methodCalls = MethodCalls()
-
-    // MARK: - Properties
-
-    var delegate: AppLockDelegate?
-
-    var isAvailable = true
-    var isActive = true
-    var isForced = false
-    var timeout: UInt = 5
-    var isLocked = false
-    var requireCustomPasscode = false
-    var isCustomPasscodeSet = false
-    var needsToNotifyUser = false
-
-    // MARK: - Methods
-
-    func deletePasscode() throws {
-        // No op
-    }
-
-    func updatePasscode(_ passcode: String) throws {
-        // No op
-    }
-
-    func beginTimer() {
-        methodCalls.beginTimer.append(())
-    }
-
-    func open() throws {
-        // No op
-    }
-
-    func evaluateAuthentication(passcodePreference: AppLockPasscodePreference, description: String, context: LAContextProtocol, callback: @escaping (AppLockAuthenticationResult, LAContextProtocol) -> Void) {
-        fatalError("Not implemented")
-    }
-
-    func evaluateAuthentication(customPasscode: String) -> AppLockAuthenticationResult {
-        fatalError("Not implemented")
-    }
-
-    // MARK: - Types
-
-    struct MethodCalls {
-
-        var beginTimer: [Void] = []
-        
-    }
-
-}
