+++ Tests/Source/Synchronization/Transcoders/ZMConversationTranscoderTests.swift
@@ -20,26 +20,6 @@ import XCTest
 import WireRequestStrategy
 import WireDataModel
 
-extension ZMConversation {
-    @objc var isFullyMuted: Bool {
-        get {
-            return mutedMessageTypes == .all
-        }
-        set {
-            mutedMessageTypes = newValue ? .all : .none
-        }
-    }
-    
-    @objc var isMutedDisplayingMentions: Bool {
-        get {
-            return mutedMessageTypes == .regular
-        }
-        set {
-            mutedMessageTypes = newValue ? .regular : .none
-        }
-    }
-}
-
 class ZMConversationTranscoderTests_Swift: ObjectTranscoderTests {
     
     var sut: ZMConversationTranscoder!
