+++ Tests/Source/UserSession/UserProfileUpdateStatusTests.swift
@@ -1139,63 +1139,3 @@ extension UserProfileUpdateStatusTests {
         XCTAssertEqual(self.sut.suggestedHandlesToCheck?.first, handle)
     }
 }
-
-// MARK: - Helpers
-class TestUserProfileUpdateObserver : NSObject, UserProfileUpdateObserver {
-    
-    var invokedCallbacks : [WireSyncEngine.UserProfileUpdateNotificationType] = []
-    
-    func passwordUpdateRequestDidFail() {
-        invokedCallbacks.append(.passwordUpdateDidFail)
-    }
-
-    func emailUpdateDidFail(_ error: Error!) {
-        invokedCallbacks.append(.emailUpdateDidFail(error: error))
-    }
-    
-    func didSendVerificationEmail() {
-        invokedCallbacks.append(.emailDidSendVerification)
-    }
-    
-    func phoneNumberVerificationCodeRequestDidFail(_ error: Error!) {
-        invokedCallbacks.append(.phoneNumberVerificationCodeRequestDidFail(error: error))
-    }
-    
-    func phoneNumberVerificationCodeRequestDidSucceed() {
-        invokedCallbacks.append(.phoneNumberVerificationCodeRequestDidSucceed)
-    }
-    
-    func phoneNumberChangeDidFail(_ error: Error!) {
-        invokedCallbacks.append(.phoneNumberChangeDidFail(error: error))
-    }
-    
-    func didCheckAvailiabilityOfHandle(handle: String, available: Bool) {
-        invokedCallbacks.append(.didCheckAvailabilityOfHandle(handle: handle, available: available))
-    }
-    
-    func didFailToCheckAvailabilityOfHandle(handle: String) {
-        invokedCallbacks.append(.didFailToCheckAvailabilityOfHandle(handle: handle))
-    }
-    
-    func didFailToSetHandleBecauseExisting() {
-        invokedCallbacks.append(.didFailToSetHandleBecauseExisting)
-    }
-    
-    func didFailToSetHandle() {
-        invokedCallbacks.append(.didFailToSetHandle)
-    }
-    
-    func didSetHandle() {
-        invokedCallbacks.append(.didSetHandle)
-    }
-    
-    func didFindHandleSuggestion(handle: String) {
-        invokedCallbacks.append(.didFindHandleSuggestion(handle: handle))
-    }
-    
-    func clearReceivedCallbacks() {
-        self.invokedCallbacks = []
-    }
-    
-}
-
