+++ Tests/Source/UserSession/ZMUserSessionTests.swift
@@ -20,29 +20,6 @@
 import Foundation
 import WireSyncEngine
 
-extension ZMUserSession {
-    @discardableResult func insertUnreadDotGeneratingMessageMessage(in conversation: ZMConversation) -> ZMSystemMessage {
-        let newTime = conversation.lastServerTimeStamp?.addingTimeInterval(5) ?? Date()
-        
-        let message = ZMSystemMessage(nonce: UUID(), managedObjectContext: self.managedObjectContext)
-        message.serverTimestamp = newTime
-        message.systemMessageType = .missedCall
-        conversation.lastServerTimeStamp = message.serverTimestamp
-        conversation.mutableMessages.add(message)
-        return message
-    }
-    
-    @discardableResult func insertConversationWithUnreadMessage() -> ZMConversation {
-        let conversation = ZMConversation.insertGroupConversation(session: self, participants: [])!
-        conversation.remoteIdentifier = UUID()
-        
-        self.insertUnreadDotGeneratingMessageMessage(in: conversation)
-        // then
-        XCTAssertNotNil(conversation.firstUnreadMessage)
-        return conversation
-    }
-}
-
 class ZMUserSessionSwiftTests: ZMUserSessionTestsBase {
