+++ Tests/Source/Utility/AnalyticsTests.swift
@@ -71,47 +71,3 @@ class AnalyticsTests: ZMTBaseTest {
         XCTAssertNil(context.analytics)
     }
 }
-
-// MARK: - Helpers
-struct EventWithAttributes: Equatable {
-    let event: String
-    let attributes: [String: NSObject]
-}
-
-func ==(lhs: EventWithAttributes, rhs: EventWithAttributes) -> Bool {
-    return lhs.event == rhs.event && lhs.attributes == rhs.attributes
-}
-
-final class MockAnalytics: NSObject, AnalyticsType {
-
-    var eventAttributes = [String : [String : NSObject]]()
-
-    public func setPersistedAttributes(_ attributes: [String : NSObject]?, for event: String) {
-        if let attributes = attributes {
-            eventAttributes[event] = attributes
-        } else {
-            eventAttributes.removeValue(forKey: event)
-        }
-    }
-
-    public func persistedAttributes(for event: String) -> [String : NSObject]? {
-        let value = eventAttributes[event] ?? [:]
-        return value
-    }
-    
-    @objc func tagEvent(_ event: String) {
-        taggedEvents.append(event)
-    }
-    
-    @objc func tagEvent(_ event: String, attributes: [String : NSObject]) {
-        taggedEventsWithAttributes.append(EventWithAttributes(event: event, attributes: attributes))
-    }
-    
-    @objc func upload() {
-        uploadCallCount += 1
-    }
-    
-    var taggedEvents = [String]()
-    var taggedEventsWithAttributes = [EventWithAttributes]()
-    var uploadCallCount = 0
-}
